# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /calendar-entries:
    get:
      operationId: listCalendarEntries
      tags: [calendar]
      parameters:
        - name: tenantId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of calendar entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CalendarEntry" }
    post:
      operationId: createCalendarEntry
      tags: [calendar]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CalendarEntryInput" }
      responses:
        "201":
          description: Created calendar entry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CalendarEntry" }
  /calendar-entries/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    patch:
      operationId: updateCalendarEntry
      tags: [calendar]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CalendarEntryUpdate" }
      responses:
        "200":
          description: Updated calendar entry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CalendarEntry" }
    delete:
      operationId: deleteCalendarEntry
      tags: [calendar]
      responses:
        "204":
          description: Deleted
  /calendar/autofill:
    post:
      operationId: autofillCalendar
      tags: [calendar]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CalendarAutofillInput" }
      responses:
        "200":
          description: Autofill result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CalendarAutofillResult" }

  # ---------- System state ----------
...
  /automation/rules:
    get:
      operationId: listAutomationRules
      tags: [automation]
      responses:
        "200":
          description: List of automation rules
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AutomationRule" }
    post:
      operationId: createAutomationRule
      tags: [automation]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AutomationRuleInput" }
      responses:
        "201":
          description: Created automation rule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AutomationRule" }
  /automation/rules/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    patch:
      operationId: updateAutomationRule
      tags: [automation]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AutomationRuleUpdate" }
      responses:
        "200":
          description: Updated automation rule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AutomationRule" }
    delete:
      operationId: deleteAutomationRule
      tags: [automation]
      responses:
        "204":
          description: Deleted
  /automation/settings:
    get:
      operationId: getAutomationSettings
      tags: [automation]
      responses:
        "200":
          description: User automation settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserAutomationSettings" }
    patch:
      operationId: updateAutomationSettings
      tags: [automation]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserAutomationSettingsUpdate" }
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserAutomationSettings" }
  /automation/trigger-events:
    get:
      operationId: listAutomationTriggerEvents
      tags: [automation]
      responses:
        "200":
          description: List of trigger events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AutomationTriggerEvent" }
    post:
      operationId: createAutomationTriggerEvent
      tags: [automation]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AutomationTriggerEventInput" }
      responses:
        "201":
          description: Created trigger event
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AutomationTriggerEvent" }
...
  /automation/auto-actions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    patch:
      operationId: updateAutoAction
      tags: [automation]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AutoActionUpdate" }
      responses:
        "200":
          description: Updated auto action
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AutoAction" }
...
  /automation/trigger:
    post:
      operationId: triggerAutomation
      tags: [automation]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AutomationTriggerInput" }
      responses:
        "200":
          description: Trigger result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcessResult" }
...
components:
  schemas:
...
    CalendarEntry:
      type: object
      required: [id, userId, title, date]
      properties:
        id: { type: integer }
        userId: { type: string }
        tenantId: { type: ["string", "null"], format: uuid }
        hubId: { type: ["integer", "null"] }
        title: { type: string }
        description: { type: ["string", "null"] }
        date: { type: string, format: date }
        startTime: { type: ["string", "null"] }
        endTime: { type: ["string", "null"] }
        focusDomain: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CalendarEntryInput:
      type: object
      required: [title, date]
      properties:
        tenantId: { type: string, format: uuid }
        hubId: { type: integer }
        title: { type: string, minLength: 1 }
        description: { type: string }
        date: { type: string, format: date }
        startTime: { type: string }
        endTime: { type: string }
        focusDomain: { type: string }
    CalendarEntryUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        date: { type: string, format: date }
        startTime: { type: string }
        endTime: { type: string }
        focusDomain: { type: string }
    CalendarAutofillInput:
      type: object
      required: [weekStart]
      properties:
        weekStart: { type: string, format: date }
        tenantId: { type: string, format: uuid }
    CalendarAutofillResult:
      type: object
      required: [created]
      properties:
        created:
          type: array
          items: { $ref: "#/components/schemas/CalendarEntry" }

    # ---------- System state ----------
...
    ProcessResult:
      type: object
      required: [processed]
      properties:
        processed: { type: integer }
        details: {}

    # ---------- Automation ----------
    AutomationRule:
      type: object
      required: [id, name, conditionType, actionTarget]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        conditionType: { type: string }
        conditionValue: { type: ["number", "null"] }
        actionTarget: { type: string }
        actionValue: { type: ["string", "null"] }
        conflictGroup: { type: ["string", "null"] }
        priority: { type: ["integer", "null"] }
        isActive: { type: boolean }
        requiresUserConfirmation: { type: boolean }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AutomationRuleInput:
      type: object
      required: [name, conditionType, actionTarget]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        conditionType: { type: string }
        conditionValue: { type: number }
        actionTarget: { type: string }
        actionValue: { type: string }
        conflictGroup: { type: string }
        priority: { type: integer }
        isActive: { type: boolean }
        requiresUserConfirmation: { type: boolean }
    AutomationRuleUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        conditionValue: { type: number }
        actionValue: { type: string }
        priority: { type: integer }
        isActive: { type: boolean }
        requiresUserConfirmation: { type: boolean }
    UserAutomationSettings:
      type: object
      required: [id, userId, automationEnabled]
      properties:
        id: { type: integer }
        userId: { type: string }
        tenantId: { type: ["string", "null"], format: uuid }
        automationEnabled: { type: boolean }
        enabledCategories: {}
        maxDailyActions: { type: ["integer", "null"] }
        priorityOverride: { type: ["string", "null"] }
        quietHours: {}
        notificationPreferences: {}
    UserAutomationSettingsUpdate:
      type: object
      properties:
        automationEnabled: { type: boolean }
        enabledCategories: {}
        maxDailyActions: { type: integer }
        priorityOverride: { type: string }
        quietHours: {}
        notificationPreferences: {}
    AutomationTriggerEvent:
      type: object
      required: [id, userId, triggerType, triggerSource]
      properties:
        id: { type: integer }
        userId: { type: string }
        tenantId: { type: ["string", "null"], format: uuid }
        triggerType: { type: string }
        triggerSource: { type: string }
        triggerData: {}
        processed: { type: boolean }
        triggeredAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    AutomationTriggerEventInput:
      type: object
      required: [triggerType, triggerSource]
      properties:
        triggerType: { type: string }
        triggerSource: { type: string }
        triggerData: {}
...
    AutoAction:
      type: object
      required: [id, userId, actionType, actionText, status]
      properties:
        id: { type: integer }
        userId: { type: string }
        tenantId: { type: ["string", "null"], format: uuid }
        hubId: { type: ["integer", "null"] }
        domainId: { type: ["integer", "null"] }
        actionType: { type: string }
        actionText: { type: string }
        actionDate: { type: string, format: date-time }
        priority: { type: ["integer", "null"] }
        status: { type: string }
        completedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    AutoActionUpdate:
      type: object
      properties:
        status: { type: string }
        completedAt: { type: string, format: date-time }
...
    AutomationTriggerInput:
      type: object
      required: [triggerType]
      properties:
        triggerType: { type: string }
        triggerSource: { type: string }
        triggerData: {}

    # ---------- Insights ----------
...