# lib/api-zod/src/generated/types/calendarEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CalendarEntry {
  id: number;
  userId: string;
  /** @nullable */
  tenantId?: string | null;
  /** @nullable */
  hubId?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  date: Date;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  focusDomain?: string | null;
  createdAt?: Date;
  updatedAt?: Date;
}